feat(#2147): move PacksTest under ParsingTrainTest
