--- eo-parser/src/test/java/org/eolang/parser/PacksTest.java
@@ -1,63 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2016-2023 Objectionary.com
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.eolang.parser;
-
-import org.eolang.jucs.ClasspathSource;
-import org.eolang.xax.XaxStory;
-import org.hamcrest.MatcherAssert;
-import org.hamcrest.Matchers;
-import org.junit.jupiter.api.Assumptions;
-import org.junit.jupiter.params.ParameterizedTest;
-
-/**
- * Test case for packs.
- * @since 1.0
- */
-final class PacksTest {
-
-    @ParameterizedTest
-    @ClasspathSource(value = "org/eolang/parser/packs/", glob = "**.yaml")
-    void parsesPacks(final String pack) throws Exception {
-        final CheckPack check = new CheckPack(pack);
-        if (check.skip()) {
-            Assumptions.abort(
-                String.format("%s is not ready", pack)
-            );
-        }
-        MatcherAssert.assertThat(
-            check.failures(),
-            Matchers.empty()
-        );
-    }
-
-    @ParameterizedTest
-    @ClasspathSource(value = "org/eolang/parser/xax/", glob = "**.yml")
-    void createsXaxStoryWithXslStylesheets(final String yaml) {
-        MatcherAssert.assertThat(
-            new XaxStory(yaml),
-            Matchers.is(true)
-        );
-    }
-
-}
